+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 - Allow adding dynamic asset collection files to a resource instead of only the Plugin builder
 - Make file endings for dynamic asset collections configurable ([#38](https://github.com/NiklasEi/bevy_asset_loader/issues/38))
 - Expose `DynamicAssetCollection` to give access to the Bevy `Assets<DynamicAssetCollection>` resource
+- Support loading lists of files as `Vec<HandleUntyped>` or `Vec<Handle<T>>` ([#48](https://github.com/NiklasEi/bevy_asset_loader/issues/48))
+  - Also supported as dynamic asset
 
 ## v0.10.0
 - Bump Bevy to version `0.7`